refractor purge server route
